fix(utils/tap): Moved tap utils to cli-utils * Migrated the tap utils * Fixed imports to point to cli-utils * Removed unused code
