--- src/utils/tap/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ...require('./getTapConfig'),
-  ...require('./getTapPackage'),
-}
